+++ django_email_learning/services/command_models/send_assignment_command.py
@@ -47,6 +47,7 @@ class SendAssignmentCommand(AbstractCommand):
             "link": self.link,
             "unsubscribe_link": content.course.generate_unsubscribe_link(self.email),
             "track_open_url": track_open_url,
+            **email_sender_service.organization_footer_context(content.course),
         }
         payload = render_to_string("emails/assignment.txt", context)